requirements.txt
numpy==1.24.1
torch==2.0.0+cu118
torchvision==0.15.1+cu118
diffusers==0.20.1
transformers==4.36.0
accelerate==0.23.0
cachetools==5.5.0
certifi==2024.2.2
click==8.1.7
clip
cmake==3.30.3
CoCa-pytorch==0.1.0
coloredlogs==15.0.1
cycler==0.12.1
distro==1.9.0
einops==0.7.0
ffmpy==0.4.0
future==1.0.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
gradio==3.44.2
gradio_client==0.5.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
imageio==2.35.1
importlib_metadata==7.1.0
importlib_resources==6.4.5
jieba==0.42.1
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kornia==0.6.2
lit==18.1.8
lmdb==1.3.0
loguru==0.6.0
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.3.4
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.12.0
open-clip-torch==2.16.0
openai==1.30.5
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.64
orjson==3.10.7
packaging==20.9
pandas==2.0.3
Pillow==9.5.0
protobuf==3.20.1
pydantic==2.7.2
pydantic_core==2.18.3
pyDeprecate==0.3.1
pytorch-lightning==1.5.0
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
safetensors==0.3.1
scikit-image==0.18.1
scikit-learn==1.1.3
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.2.0
tensorboard==2.18.0
tensorboardX==2.5
threadpoolctl==3.5.0
tifffile==2020.10.1
timm==0.4.12
tokenizers==0.15.2
torchmetrics==0.7.0
tqdm==4.65.0
triton==2.0.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.0
vit-pytorch==1.7.12
websockets==11.0.3
xformers==0.0.17
zipp==3.19.0
setproctitle